requirements.txt
Pillow==9.5.0
pyiqa==0.1.10
torch==1.12.0+cu113
torchinfo==1.8.0
